# Conflicts:
#	src/main/java/travility_back/travility/dto/ExpenseDTO.java
#	src/main/java/travility_back/travility/entity/Expense.java
#	src/main/java/travility_back/travility/repository/BudgetRepository.java
#	src/main/java/travility_back/travility/repository/ExpenseRepository.java

+++ src/main/java/travility_back/travility/service/AccountBookService.java
@@ -8,9 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import travility_back.travility.dto.AccountBookDTO;
-import travility_back.travility.dto.MemberDTO;
 import travility_back.travility.entity.AccountBook;
-import travility_back.travility.entity.Budget;
 import travility_back.travility.entity.Member;
 import travility_back.travility.repository.AccountBookRepository;
 import travility_back.travility.repository.MemberRepository;
@@ -19,7 +17,6 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -28,7 +25,6 @@ import java.util.stream.Collectors;
 public class AccountBookService {
 
     private final AccountBookRepository accountBookRepository;
-    private final MemberService memberService;
     private final MemberRepository memberRepository;
 
     //가계부 전체 조회
@@ -58,18 +54,6 @@ public class AccountBookService {
         return new AccountBookDTO(accountBook);
     }
 
-//    //가계부 삭제
-//    @Transactional
-//    public AccountBookDTO saveAccountBook(AccountBookDTO accountBookDTO) {
-//        AccountBook accountBook = convertToEntity(accountBookDTO);
-//        accountBook = accountBookRepository.save(accountBook);
-//        return convertToDTO(accountBook);
-//    }
-
-    public void deleteAccountBook(Long id) {
-        accountBookRepository.deleteById(id);
-    }
-
     //가계부 수정
     @Transactional
     public void updateAccountBook(Long id, String tripInfo, MultipartFile img) throws IOException {
@@ -96,8 +80,9 @@ public class AccountBookService {
             String newImgName = UUID.randomUUID().toString() + extension; //새 이미지 이름
             img.transferTo(new File(path,newImgName)); //지정된 경로를 가진 새 파일 객체 생성하여 업로드
 
+            System.out.println(originalName);
             //기존 이미지 파일 삭제
-            if(accountBook.getImgName() != null && !accountBook.getImgName().isEmpty()){
+            if(accountBook.getImgName().equals("default_image.png") == false && accountBook.getImgName() != null && !accountBook.getImgName().isEmpty()){
                 File oldImg = new File(path,accountBook.getImgName());
                 if (oldImg.exists()){
                     oldImg.delete();
@@ -115,44 +100,9 @@ public class AccountBookService {
         accountBook.setEndDate(accountBookDTO.getEndDate());
     }
 
-//    private AccountBookDTO convertToDTO(AccountBook accountBook) {
-//        AccountBookDTO dto = new AccountBookDTO();
-//        dto.setId(accountBook.getId());
-//        dto.setStartDate(accountBook.getStartDate());
-//        dto.setEndDate(accountBook.getEndDate());
-//        dto.setCountryName(accountBook.getCountryName());
-//        dto.setCountryFlag(accountBook.getCountryFlag());
-//        dto.setImgName(accountBook.getImgName());
-//        dto.setNumberOfPeople(accountBook.getNumberOfPeople());
-//        dto.setTitle(accountBook.getTitle());
-//        dto.setBudgets(accountBook.getBudgets().stream()
-//                .map(Budget::toDTO)
-//                .collect(Collectors.toList()));
-//        dto.setMember(new MemberDTO(accountBook.getMember()));
-//        return dto;
-//    }
-//
-//    private AccountBook convertToEntity(AccountBookDTO accountBookDTO) {
-//        AccountBook accountBook = new AccountBook();
-//        accountBook.setId(accountBookDTO.getId());
-//        accountBook.setStartDate(accountBookDTO.getStartDate());
-//        accountBook.setEndDate(accountBookDTO.getEndDate());
-//        accountBook.setCountryName(accountBookDTO.getCountryName());
-//        accountBook.setCountryFlag(accountBookDTO.getCountryFlag());
-//        accountBook.setNumberOfPeople(accountBookDTO.getNumberOfPeople());
-//        accountBook.setTitle(accountBookDTO.getTitle());
-//        accountBook.setImgName("default_image.png");
-//        accountBook.setBudgets(accountBookDTO.getBudgets().stream()
-//                .map(budgetDTO -> new Budget(budgetDTO, accountBook))
-//                .collect(Collectors.toList()));
-//
-//        MemberDTO memberDTO = accountBookDTO.getMember();
-//        if (memberDTO != null) {
-//            Member member = memberRepository.findById(memberDTO.getMemberId())
-//                    .orElseThrow(() -> new RuntimeException("User not found with id: " + memberDTO.getMemberId()));
-//            accountBook.setMember(member);
-//        }
-//
-//        return accountBook;
-//    }
+    //가계부 삭제
+    @Transactional
+    public void deleteAccountBook(Long id) {
+        accountBookRepository.deleteById(id);
+    }
 }
